# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend pendataan perusahaan berbasis kuesioner SE2026 Kuesioner L (UB \u0026 UMKM Bangunan Usaha)",
        "title": "Backend pendataan perusahaan SE2026-L.UB",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
...
        "/mahasiswa/cek-pendataan": {
            "post": {
                "description": "Cek apakah mahasiswa sudah terdaftar berdasarkan email dan NIK",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mahasiswa"
                ],
                "summary": "Cek status pendataan mahasiswa",
                "parameters": [
                    {
                        "description": "Email dan NIK",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CheckPendataanRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.CheckPendataanResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/mahasiswa/register": {
            "post": {
                "description": "Registrasi mahasiswa dengan email, NIK, dan PIN",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mahasiswa"
                ],
                "summary": "Registrasi mahasiswa baru",
                "parameters": [
                    {
                        "description": "Data registrasi",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.RegisterMahasiswaRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.RegisterMahasiswaResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/mahasiswa/verify-pin": {
            "post": {
                "description": "Verifikasi PIN untuk akses pendataan",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mahasiswa"
                ],
                "summary": "Verifikasi PIN mahasiswa",
                "parameters": [
                    {
                        "description": "Email, NIK, dan PIN",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.VerifyPINRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.VerifyPINResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/pendataan": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Pendataan"
                ],
                "summary": "List semua pendataan (admin)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Pendataan"
                            }
                        }
                    }
                }
            }
        },
...
        "/pendataan/form/{token}": {
            "get": {
                "tags": [
                    "Pendataan"
                ],
                "summary": "Get info perusahaan via token (publik)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token 64-char hex",
                        "name": "token",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Perusahaan"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Pendataan"
                ],
                "summary": "Submit kuesioner pendataan (publik)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token 64-char hex",
                        "name": "token",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Kuesioner data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Pendataan"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Pendataan"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/pendataan/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Pendataan"
                ],
                "summary": "Detail pendataan (admin)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Pendataan ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Pendataan"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.CheckPendataanRequest": {
            "type": "object",
            "required": [
                "email",
                "nik"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "mahasiswa@example.com"
                },
                "nik": {
                    "type": "string",
                    "example": "1234567890123456"
                }
            }
        },
        "models.CheckPendataanResponse": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "mahasiswa@example.com"
                },
                "message": {
                    "type": "string",
                    "example": "Anda sudah terdaftar"
                },
                "nik": {
                    "type": "string",
                    "example": "1234567890123456"
                },
                "registered": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "models.Pendataan": {
            "type": "object",
            "properties": {
                "b1r1": {
                    "type": "string"
                },
                "b1r10a": {
                    "type": "integer"
                },
                "b1r10b": {
                    "type": "integer"
                },
                "b1r11a": {
                    "type": "string"
                },
                "b1r11b": {
                    "type": "string"
                },
                "b1r11c": {
                    "type": "string"
                },
                "b1r11d": {
                    "type": "string"
                },
                "b1r11e": {
                    "type": "string"
                },
                "b1r11f": {
                    "type": "string"
                },
                "b1r12a": {
                    "type": "integer"
                },
                "b1r12b1": {
                    "type": "integer"
                },
                "b1r12b2": {
                    "type": "integer"
                },
                "b1r12b3": {
                    "type": "integer"
                },
                "b1r12b4": {
                    "type": "integer"
                },
                "b1r12b5": {
                    "type": "integer"
                },
                "b1r12b6": {
                    "type": "integer"
                },
                "b1r12c": {
                    "type": "integer"
                },
                "b1r13a": {
                    "type": "integer"
                },
                "b1r13b": {
                    "type": "integer"
                },
                "b1r14": {
                    "type": "integer"
                },
                "b1r15a": {
                    "type": "integer"
                },
                "b1r15b": {
                    "type": "integer"
                },
                "b1r15c": {
                    "type": "integer"
                },
                "b1r16a": {
                    "type": "integer"
                },
                "b1r16b": {
                    "type": "integer"
                },
                "b1r16c": {
                    "type": "integer"
                },
                "b1r17": {
                    "type": "integer"
                },
                "b1r18": {
                    "type": "integer"
                },
                "b1r19a": {
                    "type": "integer"
                },
                "b1r19b": {
                    "type": "integer"
                },
                "b1r2": {
                    "type": "string"
                },
                "b1r20a": {
                    "type": "integer"
                },
                "b1r20b": {
                    "type": "integer"
                },
                "b1r20c": {
                    "type": "integer"
                },
                "b1r21": {
                    "type": "integer"
                },
                "b1r22a": {
                    "type": "number"
                },
                "b1r22b": {
                    "type": "number"
                },
                "b1r22c": {
                    "type": "number"
                },
                "b1r22d": {
                    "type": "number"
                },
                "b1r22e": {
                    "type": "number"
                },
                "b1r22f": {
                    "type": "number"
                },
                "b1r23a": {
                    "type": "number"
                },
                "b1r23b": {
                    "type": "number"
                },
                "b1r23c": {
                    "type": "number"
                },
                "b1r23d": {
                    "type": "number"
                },
                "b1r24a": {
                    "type": "number"
                },
                "b1r24b": {
                    "type": "number"
                },
                "b1r24c": {
                    "type": "number"
                },
                "b1r24c1": {
                    "type": "integer"
                },
                "b1r24d": {
                    "type": "number"
                },
                "b1r25a": {
                    "type": "number"
                },
                "b1r25b": {
                    "type": "number"
                },
                "b1r25c": {
                    "type": "number"
                },
                "b1r25d": {
                    "type": "number"
                },
                "b1r25e": {
                    "type": "number"
                },
                "b1r25f": {
                    "type": "number"
                },
                "b1r25g": {
                    "type": "number"
                },
                "b1r3": {
                    "type": "string"
                },
                "b1r4": {
                    "type": "string"
                },
                "b1r5a": {
                    "type": "string"
                },
                "b1r5b": {
                    "type": "string"
                },
                "b1r5cAlamat": {
                    "type": "string"
                },
                "b1r5cEmail": {
                    "type": "string"
                },
                "b1r5cHomepage": {
                    "type": "string"
                },
                "b1r5cHp": {
                    "type": "string"
                },
                "b1r5cKodePos": {
                    "type": "string"
                },
                "b1r5cRt": {
                    "type": "string"
                },
                "b1r5cRw": {
                    "type": "string"
                },
                "b1r5cTelEkstensi": {
                    "type": "string"
                },
                "b1r5cTelKodeArea": {
                    "type": "string"
                },
                "b1r5cTelNomor": {
                    "type": "string"
                },
                "b1r5d": {
                    "type": "integer"
                },
                "b1r5e": {
                    "type": "string"
                },
                "b1r6a": {
                    "type": "integer"
                },
                "b1r6b": {
                    "type": "string"
                },
                "b1r6c": {
                    "type": "integer"
                },
                "b1r6cLainnya": {
                    "type": "string"
                },
                "b1r7a": {
                    "type": "integer"
                },
                "b1r7b": {
                    "type": "integer"
                },
                "b1r7c": {
                    "type": "integer"
                },
                "b1r7d": {
                    "type": "integer"
                },
                "b1r8a": {
                    "type": "string"
                },
                "b1r8b": {
                    "type": "integer"
                },
                "b1r8c": {
                    "type": "integer"
                },
                "b1r8d": {
                    "type": "string"
                },
                "b1r9a": {
                    "type": "string"
                },
                "b1r9b1": {
                    "type": "integer"
                },
                "b1r9b2": {
                    "type": "integer"
                },
                "b1r9b3": {
                    "type": "integer"
                },
                "b1r9b4": {
                    "type": "integer"
                },
                "b1r9c": {
                    "type": "integer"
                },
                "b1r9d": {
                    "type": "string"
                },
                "b1r9e": {
                    "type": "string"
                },
                "b1r9f": {
                    "type": "string"
                },
                "b1r9g": {
                    "type": "string"
                },
                "b1r9h": {
                    "type": "string"
                },
                "b1r9i": {
                    "type": "integer"
                },
                "b2catatan": {
                    "type": "string"
                },
                "b3r1": {
                    "type": "string"
                },
                "b3r3": {
                    "type": "string"
                },
                "b3r4": {
                    "type": "string"
                },
                "catatanNgibar": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "perusahaan": {
                    "$ref": "#/definitions/models.Perusahaan"
                },
                "perusahaanId": {
                    "type": "integer"
                },
                "submittedAt": {
                    "type": "string"
                },
                "tanggalPelaksanaan": {
                    "type": "string"
                },
                "ttd": {
                    "type": "boolean"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "models.Perusahaan": {
            "type": "object",
            "properties": {
                "alamat": {
                    "type": "string"
                },
                "b1r1": {
                    "type": "string"
                },
                "b1r2": {
                    "type": "string"
                },
                "b1r3": {
                    "type": "string"
                },
                "b1r4": {
                    "type": "string"
                },
                "contactPerson": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "nama": {
                    "type": "string"
                },
                "pendataan": {
                    "$ref": "#/definitions/models.Pendataan"
                },
                "pendataanUmkm": {
                    "$ref": "#/definitions/models.PendataanUmkm"
                },
                "phone": {
                    "type": "string"
                },
                "skalaUsaha": {
                    "$ref": "#/definitions/models.SkalaUsaha"
                },
                "status": {
                    "$ref": "#/definitions/models.StatusPendataan"
                },
                "token": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "models.RegisterMahasiswaRequest": {
            "type": "object",
            "required": [
                "email",
                "nik",
                "pin",
                "pinConfirm"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "mahasiswa@example.com"
                },
                "nik": {
                    "type": "string",
                    "example": "1234567890123456"
                },
                "pin": {
                    "type": "string",
                    "maxLength": 6,
                    "minLength": 6,
                    "example": "123456"
                },
                "pinConfirm": {
                    "type": "string",
                    "maxLength": 6,
                    "minLength": 6,
                    "example": "123456"
                }
            }
        },
        "models.RegisterMahasiswaResponse": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "mahasiswa@example.com"
                },
                "message": {
                    "type": "string",
                    "example": "Registrasi berhasil"
                },
                "nik": {
                    "type": "string",
                    "example": "1234567890123456"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "models.SkalaUsaha": {
            "type": "string",
            "enum": [
                "UB",
                "UMKM"
            ],
            "x-enum-varnames": [
                "SkalaUB",
                "SkalaUMKM"
            ]
        },
        "models.StatusPendataan": {
            "type": "string",
            "enum": [
                "BELUM",
                "SELESAI"
            ],
            "x-enum-varnames": [
                "StatusBelum",
                "StatusSelesai"
            ]
        },
        "models.VerifyPINRequest": {
            "type": "object",
            "required": [
                "email",
                "nik",
                "pin"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "mahasiswa@example.com"
                },
                "nik": {
                    "type": "string",
                    "example": "1234567890123456"
                },
                "pin": {
                    "type": "string",
                    "maxLength": 6,
                    "minLength": 6,
                    "example": "123456"
                }
            }
        },
        "models.VerifyPINResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "PIN valid"
                },
                "valid": {
                    "type": "boolean",
                    "example": true
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter the token with the `Bearer ` prefix, e.g. \"Bearer abcde12345\".",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}